pyproject.toml
[tool.poetry]
name = "translation-service"
version = "0.1.0"
description = ""
authors = ["Tim Koval <timkoval00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
httpx = "^0.27.0"
motor = "^3.4.0"
python-dotenv = "^1.0.1"
google-generativeai = "^0.5.4"
fastapi-pagination = "^0.12.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
